requirements.txt
tensorflow
Pillow
numpy
pandas
matplotlib
streamlit
streamlit-drawable-canvas
keras
opencv-python
# Install pyserial in VS Code
 pyserial
 pyfirmata
opencv-python
